refactor: separated api hooks from general in tracks feature
