+++ src/features/tracks/components/delete-many-tracks-dialog.tsx
@@ -2,7 +2,7 @@ import { TrashIcon } from 'lucide-react';
 
 import ConfirmationDialog from '@/components/confirmation-dialog.tsx';
 import { Button } from '@/components/ui/button.tsx';
-import useDeleteTracksMutation from '@/features/tracks/hooks/use-delete-tracks-mutation.ts';
+import useDeleteTracksMutation from '@/features/tracks/hooks/mutations/use-delete-tracks-mutation';
 import useTracksStore from '@/features/tracks/store/use-tracks-store.tsx';
 
 const DeleteManyTracksDialog = () => {
