+++ src/features/tracks/components/track-audio.tsx
@@ -5,7 +5,7 @@ import PlaybackControls from '@/components/audio/playback-controls';
 import TimeScrubber from '@/components/audio/time-scrubber';
 import { Button } from '@/components/ui/button';
 import { Card, CardContent } from '@/components/ui/card';
-import useRemoveTrackMutation from '@/features/tracks/hooks/use-remove-track-mutation.ts';
+import useRemoveTrackMutation from '@/features/tracks/hooks/mutations/use-remove-track-mutation';
 import { getAudioFile } from '@/features/tracks/lib/utils.ts';
 import useAudioStore from '@/store/use-audio-store';
 import type { Track } from '@/types/entities/track.ts';
