+++ src/features/tracks/components/track-upload.tsx
@@ -1,7 +1,7 @@
 import { UploadIcon } from 'lucide-react';
 
 import { Button } from '@/components/ui/button.tsx';
-import useTrackUploadMutation from '@/features/tracks/hooks/use-track-upload-mutation.ts';
+import useTrackUploadMutation from '@/features/tracks/hooks/mutations/use-track-upload-mutation';
 import {
   ALLOWED_AUDIO_TYPES,
   MAX_FILE_SIZE,
