+++ src/features/tracks/components/tracks-list.tsx
@@ -1,7 +1,7 @@
 import { Card, CardTitle } from '@/components/ui/card.tsx';
 import TrackCard from '@/features/tracks/components/track-card.tsx';
 import TrackCardSkeleton from '@/features/tracks/components/track-card-skeleton.tsx';
-import useTracksQuery from '@/features/tracks/hooks/use-tracks-query.ts';
+import useTracksQuery from '@/features/tracks/hooks/queries/use-tracks-query';
 import { MAX_TRACKS_PER_PAGE } from '@/features/tracks/lib/constants.ts';
 
 const TracksList = () => {
