+++ src/features/tracks/hooks/mutations/use-remove-track-mutation.ts
